requirements.txt
langchain
ipykernel
python-dotenv
langchain-community
pypdf
bs4
# arxiv
pymupdf
# wikipedia
langchain-text-splitters
# langchain_openai
# chromadb
sentence_transformers
langchain_huggingface
streamlit
langchain_groq
fastapi
langserve
huggingface_hub
sse_starlette

# langchain_chroma